✨ feat(server): refactor SMS gateway architecture and enhance modularity
- Decoupled core functionalities into distinct modules: routes, controllers, services, and models
- Centralized environment configuration into a reusable module
- Streamlined `/send-sms` and `/sms-status/:messageId` endpoints using controllers
- Improved maintainability with clear separation of concerns

Simplifies scaling and future enhancements of SMS gateway components.
